compgene/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "compgene"
version = "0.1.0"
description = "Comparative genomics pipeline for ortholog analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "CompGene Team" },
]
keywords = [
    "bioinformatics",
    "comparative-genomics",
    "orthology",
    "snakemake",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "snakemake==9.14.8",  # Strict version per Architecture spec
    "pyyaml>=6.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/compgene/compgene"
Repository = "https://github.com/compgene/compgene"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "W",    # pycodestyle warnings
]

[tool.ruff.lint.isort]
known-first-party = ["compgene"]

[tool.pytest.ini_options]
testpaths = ["tests", "workflow/adapters"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["workflow"]
omit = ["*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
